// resources/frontend/src/types/settings.type.ts
export type SettingsType = {
  address: { setting_value: string , setting_json: null }
  addressOnMap: { setting_value: string, setting_json: null }
  businessHours: { setting_value: string, setting_json: null }
  companyName: { setting_value: string, setting_json: null }
  email: { setting_value: string, setting_json: null }
  fax: { setting_value: string, setting_json: null }
  logo: { setting_value: string, setting_json: null }
  metaDesc: { setting_value: string, setting_json: null }
  metaTitle: { setting_value: string, setting_json: null }
  phone: { setting_value: string, setting_json: null }
}